awsDjango/requirements.txt
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.2
bcrypt==3.2.0
blessed==1.18.1
bootstrap4==0.1.0
botocore==1.21.40
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.3
cryptography==3.4.8
Django==3.2.7
django-autoslug==1.9.8
django-crispy-forms==1.12.0
django-emoji==2.2.2
django-emoji-picker==0.0.6
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-notify-x==0.1.9
django-tinymce==3.3.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==3.2
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
nodejs==0.1.1
optional-django==0.1.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.3.2
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
swapper==1.1.2.post1
termcolor==1.1.0
texttable==1.6.4
urllib3==1.26.6
wcwidth==0.1.9
websocket-client==0.59.0
